RF: Move create_tree* into datalad.utils
