+++ datalad/tests/utils.py
@@ -108,88 +108,6 @@ def skip_if_url_is_not_available(url, regex=None):
         raise SkipTest("%s failed to download" % url)
 
 
-# TODO: eventually we might want to make use of attr module
-class File(object):
-    """Helper for a file entry in the create_tree/@with_tree
-
-    It allows to define additional settings for entries
-    """
-    def __init__(self, name, executable=False):
-        """
-
-        Parameters
-        ----------
-        name : str
-          Name of the file
-        executable: bool, optional
-          Make it executable
-        """
-        self.name = name
-        self.executable = executable
-
-    def __str__(self):
-        return self.name
-
-
-def create_tree_archive(path, name, load, overwrite=False, archives_leading_dir=True):
-    """Given an archive `name`, create under `path` with specified `load` tree
-    """
-    from ..support.archives import compress_files
-    dirname = file_basename(name)
-    full_dirname = opj(path, dirname)
-    os.makedirs(full_dirname)
-    create_tree(full_dirname, load, archives_leading_dir=archives_leading_dir)
-    # create archive
-    if archives_leading_dir:
-        compress_files([dirname], name, path=path, overwrite=overwrite)
-    else:
-        compress_files(list(map(basename, glob.glob(opj(full_dirname, '*')))),
-                       opj(pardir, name),
-                       path=opj(path, dirname),
-                       overwrite=overwrite)
-    # remove original tree
-    shutil.rmtree(full_dirname)
-
-
-def create_tree(path, tree, archives_leading_dir=True):
-    """Given a list of tuples (name, load) create such a tree
-
-    if load is a tuple itself -- that would create either a subtree or an archive
-    with that content and place it into the tree if name ends with .tar.gz
-    """
-    lgr.log(5, "Creating a tree under %s", path)
-    if not exists(path):
-        os.makedirs(path)
-
-    if isinstance(tree, dict):
-        tree = tree.items()
-
-    for file_, load in tree:
-        if isinstance(file_, File):
-            executable = file_.executable
-            name = file_.name
-        else:
-            executable = False
-            name = file_
-        full_name = opj(path, name)
-        if isinstance(load, (tuple, list, dict)):
-            if name.endswith('.tar.gz') or name.endswith('.tar') or name.endswith('.zip'):
-                create_tree_archive(path, name, load, archives_leading_dir=archives_leading_dir)
-            else:
-                create_tree(full_name, load, archives_leading_dir=archives_leading_dir)
-        else:
-            if PY2:
-                open_kwargs = {'mode': "w"}
-                if isinstance(load, text_type):
-                    load = load.encode('utf-8')
-            else:
-                open_kwargs = {'mode': "w", 'encoding': "utf-8"}
-
-            with open(full_name, **open_kwargs) as f:
-                f.write(load)
-        if executable:
-            os.chmod(full_name, os.stat(full_name).st_mode | stat.S_IEXEC)
-
 #
 # Addition "checkers"
 #
